refactoring of rule form to enable re-use
