+++ app/javascript/packs/rules.js
@@ -3,7 +3,7 @@ import Vue from 'vue'
 import {
     ButtonPlugin, BreadcrumbPlugin, FormPlugin, FormGroupPlugin, FormInputPlugin,
     FormTextareaPlugin, BadgePlugin, CollapsePlugin, FormSelectPlugin, FormCheckboxPlugin,
-    TooltipPlugin
+    TooltipPlugin, ModalPlugin
 } from 'bootstrap-vue'
 import Rules from '../components/rules/Rules.vue'
 import RulesCodeEditorView from '../components/rules/RulesCodeEditorView.vue'
@@ -12,6 +12,11 @@ import RuleHistories from '../components/rules/RuleHistories.vue'
 import RuleNavigator from '../components/rules/RuleNavigator.vue'
 import RuleEditorHeader from '../components/rules/RuleEditorHeader.vue'
 import RuleEditor from '../components/rules/RuleEditor.vue'
+import RuleRevertModal from '../components/rules/RuleRevertModal.vue'
+import RuleForm from '../components/rules/forms/RuleForm.vue'
+import RuleDescriptionForm from '../components/rules/forms/RuleDescriptionForm.vue'
+import DisaRuleDescriptionForm from '../components/rules/forms/DisaRuleDescriptionForm.vue'
+import CheckForm from '../components/rules/forms/CheckForm.vue'
 
 Vue.use(TurbolinksAdapter)
 Vue.use(ButtonPlugin)
@@ -25,6 +30,7 @@ Vue.use(FormInputPlugin)
 Vue.use(FormSelectPlugin)
 Vue.use(FormCheckboxPlugin)
 Vue.use(TooltipPlugin)
+Vue.use(ModalPlugin)
 
 Vue.component('Rules', Rules)
 Vue.component('RulesCodeEditorView', RulesCodeEditorView)
@@ -33,6 +39,11 @@ Vue.component('RuleNavigator', RuleNavigator)
 Vue.component('RuleHistories', RuleHistories)
 Vue.component('RuleEditorHeader', RuleEditorHeader)
 Vue.component('RuleEditor', RuleEditor)
+Vue.component('RuleRevertModal', RuleRevertModal)
+Vue.component('RuleForm', RuleForm)
+Vue.component('RuleDescriptionForm', RuleDescriptionForm)
+Vue.component('DisaRuleDescriptionForm', DisaRuleDescriptionForm)
+Vue.component('CheckForm', CheckForm)
 
 document.addEventListener('turbolinks:load', () => {
     new Vue({
